Handle stack reallocation in callbacks
This patch was submitted by Father Chrysostomos in
https://rt.cpan.org/Ticket/Display.html?id=92570

ST expands to something referencing PL_stack_base. Since the order of
evalution of a=b is undefined in C, the value of PL_stack_base may be
read before sv_clone is called. sv_clone may reallocate the stack, so
the value gets written to freed memory and clone() returns its argument
instead of the clone.